adjusted routing and redirects
