+++ app/portal/courses/detail/page.module.scss
@@ -1,0 +1,6 @@
+@import "styles/variables";
+@import "styles/media";
+
+.courses {
+  display: block;
+}
