+++ app/portal/courses/page.module.scss
@@ -1,6 +1,23 @@
 @import "styles/variables";
 @import "styles/media";
 
-.courses {
-  display: block;
+section.courses {
+  .course-list {
+    display: flex;
+    flex-direction: column;
+    align-items: stretch;
+    align-content: flex-start;
+    margin-bottom: 2rem;
+    max-width: 600px;
+    border-top: 1px solid $cyd-gray;
+
+    .course-row {
+      padding: 1rem;
+      color: $cyd-pink;
+      border-bottom: 1px solid $cyd-gray;
+    }
+    .course-row:hover {
+      background-color: $cyd-light-gray;
+    }
+  }
 }
