+++ components/layout/SiteLogo/SiteLogo.jsx
@@ -3,7 +3,7 @@ import Image from "next/image"
 
 export default function SiteLogo() {
   return (
-    <Link href="/" className="no-text">
+    <Link href="/portal" className="no-text">
       <Image
         src="/images/CYD-Logo.png"
         alt="Code Your Dreams logo"
